second

+++ src/redux/contacts/operations.js
@@ -6,7 +6,11 @@ export const fetchContacts = createAsyncThunk(
   "contacts/fetchAll",
   async (_, thunkAPI) => {
     try {
-      const token = localStorage.getItem("token");
+      const state = thunkAPI.getState();
+      const token = state.auth.token;
+      if (!token) {
+        return thunkAPI.rejectWithValue("No token found");
+      }
       setAuthHeader(`Bearer ${token}`);
       const response = await api.get("/contacts");
       console.log(response);
@@ -22,7 +26,11 @@ export const addContact = createAsyncThunk(
   "contacts/addContact",
   async (contact, thunkAPI) => {
     try {
-      const token = localStorage.getItem("token");
+      const state = thunkAPI.getState();
+      const token = state.auth.token;
+      if (!token) {
+        return thunkAPI.rejectWithValue("No token found");
+      }
       setAuthHeader(`Bearer ${token}`);
       const response = await api.post("/contacts", contact);
 
@@ -37,7 +45,11 @@ export const deleteContact = createAsyncThunk(
   "contacts/deleteContact",
   async (contactID, thunkAPI) => {
     try {
-      const token = localStorage.getItem("token");
+      const state = thunkAPI.getState();
+      const token = state.auth.token;
+      if (!token) {
+        return thunkAPI.rejectWithValue("No token found");
+      }
       setAuthHeader(`Bearer ${token}`);
       const response = await api.delete(`/contacts/${contactID}`);
 
@@ -52,7 +64,11 @@ export const editContact = createAsyncThunk(
   "contacts/editContact",
   async ({ contactID, Data }, thunkAPI) => {
     try {
-      const token = localStorage.getItem("token");
+      const state = thunkAPI.getState();
+      const token = state.auth.token;
+      if (!token) {
+        return thunkAPI.rejectWithValue("No token found");
+      }
       setAuthHeader(`Bearer ${token}`);
       const response = await api.patch(`/contacts/${contactID}`, Data);
